==> code/01_scraping.ipynb <==
import pandas as pd
import praw
import os
from dotenv import load_dotenv
from datetime import datetime


# Attempts to load environment variables from .env file
# Returns True if successful, False if .env file is not found or there's an error
success = load_dotenv()

# Prints whether the environment variables were successfully loaded
# This can help with debugging if there are any issues with your 
# .env file configuration
print(f"Load successful: {success}")

# Initializing PRAW (Python Reddit API Wrapper)
reddit = praw.Reddit(
    client_id=os.getenv('REDDIT_CLIENT_ID'),
    client_secret=os.getenv('REDDIT_CLIENT_SECRET'),
    user_agent=os.getenv('REDDIT_USER_AGENT')
)

print(f"Read-only: {reddit.read_only}")

def fetch_subreddit_data(reddit, subreddit_name, post_limit=1000):
    subreddit = reddit.subreddit(subreddit_name)
    posts = []
    for post in subreddit.new(limit=post_limit):
        # Convert Unix timestamp to datetime
        created_date = datetime.fromtimestamp(post.created)
        
        posts.append({
            "title": post.title,
            "score": post.score,
            "id": post.id,
            "url": post.url,
            "comms_num": post.num_comments,
            "created": created_date,  # Now returns datetime object
            "body": post.selftext
        })
    return pd.DataFrame(posts)
post_limit = 1000

# The `subreddit_list` variable will be assigned a list of your
# two chosen subreddits for Project #3
# Example:
file_date = datetime.now().strftime('%Y%m%d')
subreddit_list = ["Airbus","Boeing"]
subreddit_list = []
subreddit_list = ["wine","beer"]

for sub in subreddit_list:
    print(f'Pulling subreddit {sub}')
    data = fetch_subreddit_data(reddit, sub, post_limit)
    data.to_csv(f'../data/subreddit_{sub}_data-{file_date}.csv', index=False, encoding='utf-8-sig')
    print(f"Fetched {len(data)} posts from {sub}")
