==> code/02_cleaning.ipynb <==
#### Load modules

import pandas as pd
from sklearn.feature_extraction.text import TfidfVectorizer

#### Load CSV data for each subreddit

wine_df = pd.DataFrame()
beer_df = pd.DataFrame()
for date in ['all']:
    wine_df = pd.concat([wine_df, pd.read_csv(f'../data/subreddit_wine_data-{date}.csv', encoding='utf-8-sig')])
    beer_df = pd.concat([beer_df, pd.read_csv(f'../data/subreddit_beer_data-{date}.csv', encoding='utf-8-sig')])

wine_df.shape, beer_df.shape

#### Combine both DataFrame

# set subreddit feature for each df
wine_df['subreddit'] = 'wine'
beer_df['subreddit'] = 'beer'

# concatenate both df
combine_df = pd.concat([wine_df, beer_df])

# combine title and body into title_body
combine_df['title_body'] = combine_df['title'] + ' ' + combine_df['body']

# drop duplicate from merging
combine_df.drop_duplicates(subset=['id'], inplace=True)

# drop row with null values
combine_df.dropna(inplace=True)

# check for null value rows
combine_df['title_body'].isnull().sum()

# display sample records
combine_df.head()

#### Export CSV data

# display DataFrame shape
combine_df.shape

# write data to CSV file
combine_df[['id','url','comms_num','created','subreddit','title_body']] \
.to_csv('../data/subreddit_combine_title_body.csv', index=False)
