==> code/03_model_selection.ipynb <==
#### Import modules

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.pipeline import Pipeline
from sklearn.metrics import confusion_matrix, ConfusionMatrixDisplay

from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.metrics import classification_report

from nltk.tokenize import sent_tokenize, word_tokenize, RegexpTokenizer
from nltk.stem import WordNetLemmatizer
from nltk.stem.porter import PorterStemmer
from nltk.corpus import stopwords

lemmatizer = WordNetLemmatizer()
p_stemmer = PorterStemmer()

##### Load data

combine_df = pd.read_csv('../data/subreddit_combine_title_body.csv')

combine_df.head()

# remove word beer or wine from posts data
combine_df['title_body'] = combine_df['title_body'].str.replace(r'beer|wine','', regex=True, case=False)

#### Vertorization

# defined base word helper function
lemma_token = WordNetLemmatizer()
stem_token = PorterStemmer()

def lemma_tokenizer(words):
    return ' '.join([lemma_token.lemmatize(w) for w in words.split()])

def stem_tokenizer(words):
    return ' '.join([stem_token.stem(w) for w in words.split()])

# setup features
X = combine_df['title_body']
y = combine_df['subreddit']

# find baseline
y.value_counts(normalize=True).mul(100).round(2)

# train / test split with Lemmatizer
X_lemma = X.apply(lemma_tokenizer)
X_lemma_train, X_lemma_test, y_lemma_train, y_lemma_test = \
train_test_split(X_lemma, y, stratify=y, random_state=42, train_size=0.8, test_size=0.2)

# train / test split with Porter stemmer
X_stem = X.apply(stem_tokenizer)
X_stem_train, X_stem_test, y_stem_train, y_stem_test = \
train_test_split(X_stem, y, stratify=y, random_state=42, train_size=0.8, test_size=0.2)

#### Training models

# defined models and parameters

vectors = {
            'CountVectorizer': {'pipeline': ('vec', CountVectorizer()),
                                'grid_params': {'vec__stop_words': [None, 'english'],
                                           'vec__min_df': [2, 3, 5],
                                           'vec__ngram_range': [(1,1), (1,2)]
                                          }
                               },
            'TfidfVectorizer': {'pipeline': ('vec', TfidfVectorizer()),
                                'grid_params': {'vec__stop_words': [None, 'english'],
                                           'vec__min_df': [2, 3, 5],
                                           'vec__ngram_range': [(1,1), (1,2)]
                                          }
                               },
}

models = {'Naive Bayes': {'pipeline': ('nb',MultinomialNB()),
                          'grid_params': {
                                          'nb__alpha': [0.001, 0.05, 0.1, 1]
                                         }
        },
        'Logistic Regression': 
                          {'pipeline': ('lr',LogisticRegression()),
                           'grid_params': {
                                          'lr__C': [0.001, 0.05, 0.1, 1],
                                          'lr__solver': ['liblinear']
                                         }
        },
        'RandomForest Classifier': 
                          {'pipeline': ('rfc', RandomForestClassifier()),
                          'grid_params': {
                                          'rfc__min_samples_split': [2 ,4, 8, 10],
                                          'rfc__min_samples_leaf': [2 ,4, 8, 10],
                                          'rfc__max_depth': [None, 1, 5 ,10, 15]
                                         }
        },
}

datas = {
    'WordNetLemmatizer': {'X_train': X_lemma_train, 'X_test': X_lemma_test, 
                          'y_train': y_lemma_train, 'y_test': y_lemma_test},
    'PorterStemmer' :    {'X_train': X_stem_train, 'X_test': X_stem_test, 
                          'y_train': y_stem_train, 'y_test': y_stem_test}    
}

# combine pipeline and gridsearch params
vec_models = {}
for vec_name, vector in vectors.items():
    for cls_name, model in models.items():
        vec_models[(vec_name, cls_name)] = {
            'pipeline':   Pipeline([
                vector['pipeline'],
                model['pipeline'],
            ]),
            'grid_params': {
                **vector['grid_params'],
                **model['grid_params'],
            }
        }

# run gridsearch
out_models = {}
i = 1
# initialize DataFrame to store score output 
model_df = pd.DataFrame(columns=['Word_Normalizer', 'Vertorizer_name', 'Model_Name', 'CV_Score', 'Training_Score', 'Testing_Score'])
for group_name, data in datas.items():
    X_train, X_test, y_train, y_test = data['X_train'], data['X_test'], data['y_train'], data['y_test']
    for name, model in vec_models.items():
        vec_name, cls_name = name
        key = (vec_name, cls_name, group_name)
        print(f"{i}. Running gridsearch for ({vec_name}, {cls_name}, {group_name})")
        out_models[key] = GridSearchCV(estimator=model['pipeline'], param_grid=model['grid_params'], \
                            scoring='accuracy', n_jobs=10, cv=2)
        out_models[key].fit(X_train, y_train)
        
        # add store for model
        model_df.loc[i] = [group_name, vec_name, cls_name, out_models[key].best_score_, 
                            out_models[key].score(X_train, y_train), out_models[key].score(X_test, y_test)]
        
        # display model result
        print(f"Best parameters is : {out_models[key].best_params_}")
        print(f"Best cross validation score is : {out_models[key].best_score_:.8f}")
        print(f"Best training score is : {out_models[key].score(X_train, y_train):.8f}")
        print(f"Best testing score is : {out_models[key].score(X_test, y_test):.8f}")
        print('-' * 80)
        i += 1

# display model scores
model_df['Score_Difference'] = model_df['Training_Score'] - model_df['Testing_Score']
model_df.sort_values(by='CV_Score', ascending=False)

#### Analyze models performance

We use `PorterStemmer` for word normalization and `CountVectorizer` for feature extraction with `Naive Bayes` (model #7).
- Best parameters is : {'nb__alpha': 0.1, 'vec__min_df': 2, 'vec__ngram_range': (1, 1), 'vec__stop_words': None}

While the CV score is nearly identical to the best model #1, the low difference between training and testing scores suggests it is not overfitting, unlike model #1.

# select model
model = out_models[('CountVectorizer', 'Naive Bayes', 'WordNetLemmatizer')].best_estimator_

# get data
data = datas['PorterStemmer']
X_train, X_test, y_train, y_test = data['X_train'], data['X_test'], data['y_train'], data['y_test']
model.fit(X_train, y_train)
y_pred = model.predict(X_test)

# get classification report
score = classification_report(y_test, y_pred, output_dict=True)
score_df = pd.DataFrame(score).T

# dispkay baseline
y.value_counts(normalize=True).round(2)
# classes are well-balanced.

# display classification report
score_df

It's a good classification model has high precision, recall, and F1-scores for each class.
