==> code/04_sentiment_analysis.ipynb <==
#### Import modules

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.feature_extraction.text import CountVectorizer

import nltk
from nltk.stem import WordNetLemmatizer
from nltk.sentiment.vader import SentimentIntensityAnalyzer

#### Load subreddit data

# load data from CSV
topic_df = pd.read_csv('../data/subreddit_combine_title_body.csv')

# display sample data
topic_df.head(2)

#### Load class mapping data

# load data from CSV
class_df = pd.read_csv('../data/wine_beer_class.csv')

# convert words to lowercase
class_df['variety'] = class_df['variety'].str.lower()

# display sample rows
class_df.head(5)

# display dataframe shape
class_df.shape

# get word list for vertorization
variety = pd.Series(class_df['variety'].unique()).str.lower()
class_words = {w: i for i, w in enumerate(variety)}

# vertorization only word from the `variety`
cvec = CountVectorizer(vocabulary=class_words, stop_words='english', ngram_range=(1, 2))
cvec_word = cvec.fit_transform(topic_df['title_body'])

# normalize words
lemma_token = WordNetLemmatizer()

# get words from vectorizer
words = [lemma_token.lemmatize(w) for w in cvec.get_feature_names_out()]

# generate word class dataframe
word_df = pd.DataFrame(cvec_word.toarray(), columns=words)

# copy id and subreddit from newly create word
word_df['id'] = topic_df['id']
word_df['subreddit'] = topic_df['subreddit']

# convert number to boolean for count only once per topic
beer_words_df = pd.DataFrame(data={'type': 'beer', 'topic_count': word_df[word_df['subreddit']=='beer'].astype(bool).sum()})
wine_words_df = pd.DataFrame(data={'type': 'wine', 'topic_count': word_df[word_df['subreddit']=='wine'].astype(bool).sum()})

# combine beer and wine words DataFrame using `word` as key
topic_words_df = pd.concat([beer_words_df, wine_words_df])

# drop unused columns
topic_words_df.drop(['id','subreddit'], inplace=True)

# copy word from the index to a new column
topic_words_df['word'] = topic_words_df.index

# reset index
topic_words_df.reset_index(drop=True, inplace=True)

topic_words_df.head()

# mapping word class to topic_word_df
def map_characteristic(word):
    topic_words_df.loc[
      (topic_words_df['type']==word['type']) \
    & (topic_words_df['word']==word['variety']), 'characteristic'] = word['characteristic']

class_df.apply(map_characteristic, axis=1)
topic_words_df.dropna(inplace=True)

# get characteristic for beer
beer_class_df = topic_words_df[topic_words_df['type']=='beer'] \
    .groupby(['characteristic','word']).sum() \
    .reset_index() \
    .sort_values(by=['characteristic','topic_count'], ascending=False)

beer_class_df.head()

# get characteristic for wine
wine_class_df = topic_words_df[topic_words_df['type']=='wine'] \
    .groupby(['characteristic','word']).sum() \
    .reset_index() \
    .sort_values(by=['characteristic','topic_count'], ascending=False)

wine_class_df.head()

all_class_df = pd.concat([beer_class_df, wine_class_df]).reset_index(drop=True)

# convert compound score to binary value
def get_sum_compound(sentiment):
    if sentiment['compound'] <= -0.05:
        return 'negative'
    else:
        return 'not_negative'

# calculate sentiment for each topic
sia = SentimentIntensityAnalyzer()
topic_class_df = topic_df[['id','subreddit','title_body']].copy()
topic_class_df['sentiment'] = topic_class_df['title_body'].apply(lambda d: sia.polarity_scores(d))

# separate sentiment value to multiple columns
topic_class_df['sent_neg'] = topic_class_df['sentiment'].apply(lambda s: s['neg'])
topic_class_df['sent_neu'] = topic_class_df['sentiment'].apply(lambda s: s['neu'])
topic_class_df['sent_pos'] = topic_class_df['sentiment'].apply(lambda s: s['pos'])
topic_class_df['sent_comp'] = topic_class_df['sentiment'].apply(lambda s: s['compound'])
topic_class_df['sum_comp'] = topic_class_df['sentiment'].apply(get_sum_compound)
topic_class_df.set_index('id', inplace=True)

# concat sentiment with word verterization dataframe
topic_class_df = pd.concat([topic_class_df, word_df.set_index('id').drop(columns='subreddit')], axis=1)
topic_class_df.dropna(inplace=True)
topic_class_df.drop(columns=['sentiment'], inplace=True)
topic_class_df.to_csv('../data/topic_word_class.csv')

# show sample data
topic_class_df[['subreddit','title_body','sent_neg','sent_neu','sent_pos','sent_comp','sum_comp']].head()

# update summarize binary sentiment for each word class
for i, word in all_class_df.iterrows():
    a = topic_class_df[(topic_class_df['subreddit']==word['type']) & topic_class_df[word['word']]]['sum_comp'].value_counts(normalize=True)
    all_class_df.loc[i, 'negative'] = a['negative'] if 'negative' in a else 0
    all_class_df.loc[i, 'not_negative'] = a['not_negative'] if 'not_negative' in a else 0

# show combine result
all_class_df.sort_values('characteristic').head()

# save result to CSV
all_class_df.dropna(inplace=True)
all_class_df.to_csv('../data/all_type_class.csv', index=False)
